fix not being able to send messages while the bot is waiting for a poll response

+++ zeus.py
@@ -63,9 +63,9 @@ async def on_message(x):
     else:
         if x.guild.id != serverid:
             return
-        if x.author.id in lrs: # if the bot is expecting a response, don't do anything with the message here
+        if x.author.id in lrs and x.channel.id == report.channel: # if the bot is expecting a response, don't do anything with the message here
             return
-        if x.author.id in rs: # check if bot needs response to either a poll or a text question
+        if x.author.id in rs and x.channel.id == reportchannel: # check if bot needs response to either a poll or a text question
             return await x.delete()
         if x.channel.id == queuechannel:
             return await x.delete()
